requirements.txt
asgiref==3.6.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.2.1
django-request==1.6.2
Flask==2.3.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.12.0
MarkupSafe==2.1.2
numpy==1.24.3
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
python-dateutil==2.8.2
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.4
